fix(core): refactor ReferenceInput to render ReferencePreview
